chore(kubernetes): tests adapted
